requirements.txt
yfinance
pandas
matplotlib
python-dotenv
requests
